@@ -1,6 +1,8 @@
 module github.com/pydio/cells/v4
 
-go 1.19
+go 1.21
+
+toolchain go1.21.4
 
 require (
 	buf.build/gen/go/bufbuild/protovalidate/protocolbuffers/go v1.31.0-20230914171853-63dfe56cc2c4.1
