@@ -51,7 +51,7 @@ type Model struct {
 	Collections []Collection
 }
 
-func (m Model) Init(ctx context.Context, dao DAO) error {
+func (m Model) Init(ctx context.Context, db *mongo.Database) error {
 	for _, col := range m.Collections {
 		opts := &options.CreateCollectionOptions{}
 		if col.DefaultCollation.Locale != "" {
@@ -61,10 +61,10 @@ func (m Model) Init(ctx context.Context, dao DAO) error {
 			}
 		}
 		name := col.Name
-		if dao.Prefix() != "" {
-			name = dao.Prefix() + "_" + name
-		}
-		if e := dao.DB().CreateCollection(ctx, name, opts); e != nil {
+		//if dao.Prefix() != "" {
+		//	name = dao.Prefix() + "_" + name
+		//}
+		if e := db.CreateCollection(ctx, name, opts); e != nil {
 			if _, ok := e.(mongo.CommandError); !ok {
 				return e
 			} else {
@@ -80,7 +80,7 @@ func (m Model) Init(ctx context.Context, dao DAO) error {
 						keys = append(keys, primitive.E{Key: key, Value: sort})
 					}
 				}
-				if _, e := dao.DB().Collection(name).Indexes().CreateOne(ctx, mongo.IndexModel{Keys: keys}); e != nil {
+				if _, e := db.Collection(name).Indexes().CreateOne(ctx, mongo.IndexModel{Keys: keys}); e != nil {
 					return e
 				} else {
 					// fmt.Println("Successfully created index " + iName)
