@@ -23,19 +23,17 @@ package grpc
 
 import (
 	"context"
-	"github.com/pydio/cells/v4/common/dao/mysql"
-	"github.com/pydio/cells/v4/common/dao/sqlite"
-	commonsql "github.com/pydio/cells/v4/common/sql"
+	"github.com/pydio/cells/v4/common/server"
+	"github.com/pydio/cells/v4/common/storage"
+	"github.com/pydio/cells/v4/common/utils/configx"
 	"github.com/pydio/cells/v4/idm/acl"
 	"google.golang.org/grpc"
 
 	"github.com/pydio/cells/v4/common"
 	"github.com/pydio/cells/v4/common/broker"
-	"github.com/pydio/cells/v4/common/nodes/meta"
 	"github.com/pydio/cells/v4/common/proto/idm"
 	"github.com/pydio/cells/v4/common/proto/tree"
 	"github.com/pydio/cells/v4/common/runtime"
-	"github.com/pydio/cells/v4/common/service"
 )
 
 const ServiceName = common.ServiceGrpcNamespace_ + common.ServiceAcl
@@ -43,70 +41,64 @@ const ServiceName = common.ServiceGrpcNamespace_ + common.ServiceAcl
 func init() {
 
 	runtime.Register("main", func(ctx context.Context) {
-		var s service.Service
+		var srv grpc.ServiceRegistrar
+		if !server.Get(&srv) {
+			panic("no grpc server available")
+		}
 
-		s = service.NewService(
-			service.Name(ServiceName),
-			service.Context(ctx),
-			service.Tag(common.ServiceTagIdm),
-			service.Description("Access Control List service"),
-			//service.WithStorage(acl.NewDAO,
-			//	service.WithStoragePrefix("idm_acl"),
-			//	service.WithStorageSupport(mysql.Driver, sqlite.Driver),
-			//),
-			service.WithTODOStorage(acl.NewDAO, commonsql.NewDAO,
-				service.WithStoragePrefix("idm_acl"),
-				service.WithStorageSupport(mysql.Driver, sqlite.Driver),
-			),
-			service.TODOMigrations(func() []*service.Migration {
-				return []*service.Migration{
-					{
-						TargetVersion: service.ValidVersion("1.2.0"),
-						Up:            UpgradeTo120,
-					},
-				}
-			}),
-			service.Metadata(meta.ServiceMetaProvider, "stream"),
-			service.WithGRPC(func(ctx context.Context, server grpc.ServiceRegistrar) error {
-				handler := NewHandler(ctx, s)
+		dao, err := acl.NewDAO(ctx, storage.Main)
+		if err != nil {
+			panic(err)
+		}
 
-				idm.RegisterACLServiceServer(server, handler)
-				tree.RegisterNodeProviderStreamerServer(server, handler)
+		opts := configx.New()
+		dao.Init(ctx, opts)
 
-				// Clean acls on Ws or Roles deletion
-				rCleaner := &WsRolesCleaner{Handler: handler}
-				if e := broker.SubscribeCancellable(ctx, common.TopicIdmEvent, func(ctx context.Context, message broker.Message) error {
-					ev := &idm.ChangeEvent{}
-					if e := message.Unmarshal(ev); e == nil {
-						return rCleaner.Handle(ctx, ev)
-					}
-					return nil
-				}); e != nil {
-					return e
-				}
+		//service.TODOMigrations(func() []*service.Migration {
+		//	return []*service.Migration{
+		//		{
+		//			TargetVersion: service.ValidVersion("1.2.0"),
+		//			Up:            UpgradeTo120,
+		//		},
+		//	}
+		//}),
+		//service.Metadata(meta.ServiceMetaProvider, "stream")
 
-				nCleaner := newNodesCleaner(ctx, handler)
-				if e := broker.SubscribeCancellable(ctx, common.TopicTreeChanges, func(ctx context.Context, message broker.Message) error {
-					ev := &tree.NodeChangeEvent{}
-					if e := message.Unmarshal(ev); e == nil {
-						return nCleaner.Handle(ctx, ev)
-					}
-					return nil
-				}); e != nil {
-					return e
-				}
+		handler := NewHandler(ctx, dao.(acl.DAO))
 
-				// For when it will be used: clean locks at startup
-				//	dao := servicecontext.GetDAO(m.Options().Context).(acl.DAO)
-				//	if dao != nil {
-				//		q, _ := anypb.New(&idm.ACLSingleQuery{Actions: []*idm.ACLAction{{Name: permissions.AclLock.Name}}})
-				//		if num, _ := dao.Del(&service2.Query{SubQueries: []*anypb.Any{q}}); num > 0 {
-				//			log.Logger(m.Options().Context).Info(fmt.Sprintf("Cleaned %d locks in ACLs", num))
-				//		}
-				//	}
+		idm.RegisterACLServiceServer(srv, handler)
+		tree.RegisterNodeProviderStreamerServer(srv, handler)
 
-				return nil
-			}),
-		)
+		// Clean acls on Ws or Roles deletion
+		rCleaner := &WsRolesCleaner{Handler: handler}
+		if e := broker.SubscribeCancellable(ctx, common.TopicIdmEvent, func(ctx context.Context, message broker.Message) error {
+			ev := &idm.ChangeEvent{}
+			if e := message.Unmarshal(ev); e == nil {
+				return rCleaner.Handle(ctx, ev)
+			}
+			return nil
+		}); e != nil {
+			panic(e)
+		}
+
+		nCleaner := newNodesCleaner(ctx, handler)
+		if e := broker.SubscribeCancellable(ctx, common.TopicTreeChanges, func(ctx context.Context, message broker.Message) error {
+			ev := &tree.NodeChangeEvent{}
+			if e := message.Unmarshal(ev); e == nil {
+				return nCleaner.Handle(ctx, ev)
+			}
+			return nil
+		}); e != nil {
+			panic(e)
+		}
+
+		// For when it will be used: clean locks at startup
+		//	dao := servicecontext.GetDAO(m.Options().Context).(acl.DAO)
+		//	if dao != nil {
+		//		q, _ := anypb.New(&idm.ACLSingleQuery{Actions: []*idm.ACLAction{{Name: permissions.AclLock.Name}}})
+		//		if num, _ := dao.Del(&service2.Query{SubQueries: []*anypb.Any{q}}); num > 0 {
+		//			log.Logger(m.Options().Context).Info(fmt.Sprintf("Cleaned %d locks in ACLs", num))
+		//		}
+		//	}
 	})
 }
