@@ -43,15 +43,14 @@ import (
 type userKeyStore struct {
 	enc.UnimplementedUserKeyStoreServer
 
-	service.Service
-	dao func(context.Context) key.DAO
+	dao key.DAO
 
 	master []byte
 	legacy []byte
 }
 
 // NewUserKeyStore creates a master password based
-func NewUserKeyStore(ctx context.Context, svc service.Service) (enc.UserKeyStoreServer, error) {
+func NewUserKeyStore(ctx context.Context, dao key.DAO) (enc.UserKeyStoreServer, error) {
 
 	// TODO, retrieve keyring from context
 	masterPasswordStr, err := service.KeyringFromContext(ctx).Get(common.ServiceGrpcNamespace_+common.ServiceUserKey, common.KeyringMasterKey)
@@ -65,9 +64,9 @@ func NewUserKeyStore(ctx context.Context, svc service.Service) (enc.UserKeyStore
 	}
 
 	return &userKeyStore{
-		Service: svc,
-		dao:     service.DAOProvider[key.DAO](svc),
-		master:  masterPassword,
+		// Service: svc,
+		dao:    dao,
+		master: masterPassword,
 	}, nil
 }
 
@@ -78,7 +77,7 @@ func (ukm *userKeyStore) AddKey(ctx context.Context, req *enc.AddKeyRequest) (*e
 		return nil, err
 	}
 
-	return &enc.AddKeyResponse{}, ukm.dao(ctx).SaveKey(req.Key)
+	return &enc.AddKeyResponse{}, ukm.dao.SaveKey(req.Key)
 }
 
 func (ukm *userKeyStore) GetKey(ctx context.Context, req *enc.GetKeyRequest) (*enc.GetKeyResponse, error) {
@@ -90,7 +89,7 @@ func (ukm *userKeyStore) GetKey(ctx context.Context, req *enc.GetKeyRequest) (*e
 
 	var err error
 	var version int
-	rsp.Key, version, err = ukm.dao(ctx).GetKey(user, req.KeyID)
+	rsp.Key, version, err = ukm.dao.GetKey(user, req.KeyID)
 	if err != nil {
 		return nil, err
 	}
@@ -116,13 +115,13 @@ func (ukm *userKeyStore) AdminListKeys(ctx context.Context, req *enc.AdminListKe
 	rsp := &enc.AdminListKeysResponse{}
 	var err error
 
-	rsp.Keys, err = ukm.dao(ctx).ListKeys(common.PydioSystemUsername)
+	rsp.Keys, err = ukm.dao.ListKeys(common.PydioSystemUsername)
 	return rsp, err
 }
 
 func (ukm *userKeyStore) AdminCreateKey(ctx context.Context, req *enc.AdminCreateKeyRequest) (*enc.AdminCreateKeyResponse, error) {
 
-	keyDao := ukm.dao(ctx)
+	keyDao := ukm.dao
 
 	rsp := &enc.AdminCreateKeyResponse{Success: true}
 
@@ -141,7 +140,7 @@ func (ukm *userKeyStore) AdminCreateKey(ctx context.Context, req *enc.AdminCreat
 
 func (ukm *userKeyStore) AdminDeleteKey(ctx context.Context, req *enc.AdminDeleteKeyRequest) (*enc.AdminDeleteKeyResponse, error) {
 
-	return &enc.AdminDeleteKeyResponse{}, ukm.dao(ctx).DeleteKey(common.PydioSystemUsername, req.KeyID)
+	return &enc.AdminDeleteKeyResponse{}, ukm.dao.DeleteKey(common.PydioSystemUsername, req.KeyID)
 }
 
 func (ukm *userKeyStore) AdminImportKey(ctx context.Context, req *enc.AdminImportKeyRequest) (*enc.AdminImportKeyResponse, error) {
@@ -151,7 +150,7 @@ func (ukm *userKeyStore) AdminImportKey(ctx context.Context, req *enc.AdminImpor
 	log.Logger(ctx).Debug("Received request", zap.Any("Data", req))
 
 	var k *enc.Key
-	k, _, err = ukm.dao(ctx).GetKey(common.PydioSystemUsername, req.Key.ID)
+	k, _, err = ukm.dao.GetKey(common.PydioSystemUsername, req.Key.ID)
 	if err != nil {
 		if errors.FromError(err).Code != 404 {
 			return nil, err
@@ -206,7 +205,7 @@ func (ukm *userKeyStore) AdminImportKey(ctx context.Context, req *enc.AdminImpor
 	})
 
 	log.Logger(ctx).Debug("Saving new key")
-	err = ukm.dao(ctx).SaveKey(req.Key)
+	err = ukm.dao.SaveKey(req.Key)
 	if err != nil {
 		rsp.Success = false
 		return rsp, errors.InternalServerError(common.ServiceEncKey, "failed to save imported key, cause: %s", err.Error())
@@ -225,7 +224,7 @@ func (ukm *userKeyStore) AdminExportKey(ctx context.Context, req *enc.AdminExpor
 
 	rsp := &enc.AdminExportKeyResponse{}
 	var version int
-	rsp.Key, version, err = ukm.dao(ctx).GetKey(common.PydioSystemUsername, req.KeyID)
+	rsp.Key, version, err = ukm.dao.GetKey(common.PydioSystemUsername, req.KeyID)
 	if err != nil {
 		return rsp, err
 	}
@@ -241,7 +240,7 @@ func (ukm *userKeyStore) AdminExportKey(ctx context.Context, req *enc.AdminExpor
 	})
 
 	// We update the key
-	err = ukm.dao(ctx).SaveKey(rsp.Key, version)
+	err = ukm.dao.SaveKey(rsp.Key, version)
 	if err != nil {
 		return rsp, errors.InternalServerError(common.ServiceEncKey, "failed to update key info, cause: %s", err.Error())
 	}
