@@ -23,6 +23,7 @@ package grpc
 import (
 	"context"
 	"fmt"
+	"google.golang.org/protobuf/types/known/anypb"
 
 	"strings"
 
@@ -34,7 +35,6 @@ import (
 	pbservice "github.com/pydio/cells/v4/common/proto/service"
 	"github.com/pydio/cells/v4/common/proto/tree"
 	"github.com/pydio/cells/v4/common/runtime"
-	"github.com/pydio/cells/v4/common/service"
 	servicecontext "github.com/pydio/cells/v4/common/service/context"
 	"github.com/pydio/cells/v4/common/service/context/metadata"
 	"github.com/pydio/cells/v4/common/utils/cache"
@@ -47,17 +47,15 @@ type Handler struct {
 	idm.UnimplementedUserMetaServiceServer
 	tree.UnimplementedNodeProviderStreamerServer
 
-	service.Service
-	dao func(context.Context) meta.DAO
+	dao meta.DAO
 
 	searchCache cache.Cache
 }
 
-func NewHandler(ctx context.Context, svc service.Service) *Handler {
+func NewHandler(ctx context.Context, dao meta.DAO) *Handler {
 	c, _ := cache.OpenCache(context.TODO(), runtime.CacheURL(common.ServiceGrpcNamespace_+common.ServiceUserMeta))
 	h := &Handler{
-		Service: svc,
-		dao:     service.DAOProvider[meta.DAO](svc),
+		dao: dao,
 	}
 	h.searchCache = c
 	go func() {
@@ -75,7 +73,7 @@ func (h *Handler) Stop() {
 func (h *Handler) UpdateUserMeta(ctx context.Context, request *idm.UpdateUserMetaRequest) (*idm.UpdateUserMetaResponse, error) {
 
 	response := &idm.UpdateUserMetaResponse{}
-	namespaces, _ := h.dao(ctx).GetNamespaceDao().List()
+	namespaces, _ := h.dao.GetNamespaceDao().List()
 	nodes := make(map[string]*tree.Node)
 	sources := make(map[string]*tree.Node)
 	for _, metaData := range request.MetaDatas {
@@ -88,7 +86,7 @@ func (h *Handler) UpdateUserMeta(ctx context.Context, request *idm.UpdateUserMet
 				return nil, fmt.Errorf("make sure to use JSON format for metadata: %s", er.Error())
 			}
 			// ADD / UPDATE
-			if newMeta, prev, err := h.dao(ctx).Set(metaData); err == nil {
+			if newMeta, prev, err := h.dao.Set(metaData); err == nil {
 				response.MetaDatas = append(response.MetaDatas, newMeta)
 				prevValue = prev
 			} else {
@@ -96,7 +94,7 @@ func (h *Handler) UpdateUserMeta(ctx context.Context, request *idm.UpdateUserMet
 			}
 		} else {
 			// DELETE
-			if prev, err := h.dao(ctx).Del(metaData); err == nil {
+			if prev, err := h.dao.Del(metaData); err == nil {
 				prevValue = prev
 			} else {
 				return nil, err
@@ -141,9 +139,27 @@ func (h *Handler) UpdateUserMeta(ctx context.Context, request *idm.UpdateUserMet
 					})
 				}
 			}
-			metas, e := h.dao(ctx).Search([]string{}, []string{target.Uuid}, "", "", &pbservice.ResourcePolicyQuery{
+			searchUserMetaAny, err := anypb.New(&idm.SearchUserMetaRequest{
+				NodeUuids: []string{target.Uuid},
+			})
+			if err != nil {
+				continue
+			}
+
+			resourceQueryAny, err := anypb.New(&pbservice.ResourcePolicyQuery{
 				Subjects: subjects,
 			})
+			if err != nil {
+				continue
+			}
+
+			query := &pbservice.Query{
+				SubQueries: []*anypb.Any{
+					searchUserMetaAny, resourceQueryAny,
+				},
+				Operation: pbservice.OperationType_AND,
+			}
+			metas, e := h.dao.Search(query)
 			if e != nil {
 				continue
 			}
@@ -167,9 +183,28 @@ func (h *Handler) UpdateUserMeta(ctx context.Context, request *idm.UpdateUserMet
 // SearchUserMeta retrieves meta based on various criteria.
 func (h *Handler) SearchUserMeta(request *idm.SearchUserMetaRequest, stream idm.UserMetaService_SearchUserMetaServer) error {
 
-	ctx := stream.Context()
+	// ctx := stream.Context()
 
-	results, err := h.dao(ctx).Search(request.MetaUuids, request.NodeUuids, request.Namespace, request.ResourceSubjectOwner, request.ResourceQuery)
+	searchUserMetaAny, err := anypb.New(request)
+	if err != nil {
+		return err
+	}
+
+	resourceQueryAny, err := anypb.New(request.ResourceQuery)
+	if err != nil {
+		return err
+	}
+
+	query := &pbservice.Query{
+		SubQueries: []*anypb.Any{
+			searchUserMetaAny, resourceQueryAny,
+		},
+		Offset:    0,
+		Limit:     10,
+		Operation: pbservice.OperationType_AND,
+	}
+
+	results, err := h.dao.Search(query)
 	if err != nil {
 		return err
 	}
@@ -210,9 +245,28 @@ func (h *Handler) ReadNodeStream(stream tree.NodeProviderStreamer_ReadNodeStream
 		if r, ok := h.resultsFromCache(node.Uuid, subjects); ok {
 			results = r
 		} else {
-			results, err = h.dao(ctx).Search([]string{}, []string{node.Uuid}, "", "", &pbservice.ResourcePolicyQuery{
+			searchUserMetaAny, err := anypb.New(&idm.SearchUserMetaRequest{
+				NodeUuids: []string{node.Uuid},
+			})
+			if err != nil {
+				return err
+			}
+
+			resourceQueryAny, err := anypb.New(&pbservice.ResourcePolicyQuery{
 				Subjects: subjects,
 			})
+			if err != nil {
+				return err
+			}
+
+			query := &pbservice.Query{
+				SubQueries: []*anypb.Any{
+					searchUserMetaAny, resourceQueryAny,
+				},
+				Operation: pbservice.OperationType_AND,
+			}
+
+			results, err = h.dao.Search(query)
 			log.Logger(ctx).Debug(fmt.Sprintf("Got %d results for node", len(results)), node.ZapUuid())
 			if err == nil {
 				h.resultsToCache(node.Uuid, subjects, results)
@@ -233,7 +287,7 @@ func (h *Handler) ReadNodeStream(stream tree.NodeProviderStreamer_ReadNodeStream
 func (h *Handler) UpdateUserMetaNamespace(ctx context.Context, request *idm.UpdateUserMetaNamespaceRequest) (*idm.UpdateUserMetaNamespaceResponse, error) {
 
 	response := &idm.UpdateUserMetaNamespaceResponse{}
-	dao := h.dao(ctx).GetNamespaceDao()
+	dao := h.dao.GetNamespaceDao()
 	for _, metaNameSpace := range request.Namespaces {
 		if err := dao.Del(metaNameSpace); err != nil {
 			return nil, err
@@ -264,9 +318,9 @@ func (h *Handler) UpdateUserMetaNamespace(ctx context.Context, request *idm.Upda
 // ListUserMetaNamespace List all namespaces from underlying DAO.
 func (h *Handler) ListUserMetaNamespace(request *idm.ListUserMetaNamespaceRequest, stream idm.UserMetaService_ListUserMetaNamespaceServer) error {
 
-	ctx := stream.Context()
+	// ctx := stream.Context()
 
-	dao := h.dao(ctx).GetNamespaceDao()
+	dao := h.dao.GetNamespaceDao()
 	if results, err := dao.List(); err == nil {
 		for _, result := range results {
 			stream.Send(&idm.ListUserMetaNamespaceResponse{UserMetaNamespace: result})
