@@ -23,14 +23,11 @@ package namespace
 
 import (
 	"context"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
-
 	"github.com/pydio/cells/v4/common/dao"
 	"github.com/pydio/cells/v4/common/proto/idm"
-	"github.com/pydio/cells/v4/common/sql"
 	"github.com/pydio/cells/v4/common/sql/resources"
+	"github.com/pydio/cells/v4/common/storage"
+	"gorm.io/gorm"
 )
 
 const (
@@ -46,25 +43,17 @@ type DAO interface {
 	List() (map[string]*idm.UserMetaNamespace, error)
 }
 
-func NewDAO(ctx context.Context, o dao.DAO) (dao.DAO, error) {
-	switch v := o.(type) {
-	case sql.DAO:
-		dialector := sqlite.Open(v.Dsn())
-		db, err := gorm.Open(dialector, &gorm.Config{
-			//DisableForeignKeyConstraintWhenMigrating: true,
-			FullSaveAssociations: true,
-			Logger:               logger.Default.LogMode(logger.Info),
-		})
-		if err != nil {
-			return nil, err
-		}
+func NewDAO(ctx context.Context, store storage.Storage) (dao.DAO, error) {
+	var db *gorm.DB
 
-		resourcesDAO, err := resources.NewDAO(ctx, o)
+	if store.Get(ctx, &db) {
+		resourcesDAO, err := resources.NewDAO(ctx, store)
 		if err != nil {
 			return nil, err
 		}
 
 		return &sqlimpl{db: db, resourcesDAO: resourcesDAO.(resources.DAO)}, nil
 	}
-	return nil, dao.UnsupportedDriver(o)
+
+	return nil, storage.UnsupportedDriver(store)
 }
