@@ -23,6 +23,8 @@ package policy
 
 import (
 	"context"
+	"github.com/pydio/cells/v4/common/storage"
+	"gorm.io/gorm"
 
 	"github.com/ory/ladon"
 
@@ -41,10 +43,12 @@ type DAO interface {
 	DeletePolicyGroup(ctx context.Context, group *idm.PolicyGroup) error
 }
 
-func NewDAO(ctx context.Context, o dao.DAO) (dao.DAO, error) {
-	switch v := o.(type) {
-	case sql.DAO:
-		return &sqlimpl{DAO: v}, nil
+func NewDAO(ctx context.Context, store storage.Storage) (dao.DAO, error) {
+	var db *gorm.DB
+
+	if store.Get(ctx, &db) {
+		return &sqlimpl{db: db}, nil
 	}
-	return nil, dao.UnsupportedDriver(o)
+
+	return nil, storage.UnsupportedDriver(store)
 }
