@@ -49,13 +49,13 @@ type Handler struct {
 	idm.UnimplementedRoleServiceServer
 
 	service.Service
-	dao func(context.Context) role.DAO
+	dao role.DAO
 }
 
-func NewHandler(ctx context.Context, svc service.Service) idm.RoleServiceServer {
+func NewHandler(ctx context.Context, svc service.Service, dao role.DAO) idm.RoleServiceServer {
 	return &Handler{
 		Service: svc,
-		dao:     service.DAOProvider[role.DAO](svc),
+		dao:     dao,
 	}
 }
 
@@ -67,7 +67,7 @@ func (h *Handler) CreateRole(ctx context.Context, req *idm.CreateRoleRequest) (*
 		return nil, errors.BadRequest("forbidden.characters", "commas are not allowed in role uuid")
 	}
 
-	r, update, err := h.dao(ctx).Add(req.Role)
+	r, update, err := h.dao.Add(req.Role)
 	if err != nil {
 		return nil, err
 	}
@@ -80,7 +80,7 @@ func (h *Handler) CreateRole(ctx context.Context, req *idm.CreateRoleRequest) (*
 		}
 	}
 
-	err = h.dao(ctx).AddPolicies(update, r.Uuid, r.Policies)
+	err = h.dao.AddPolicies(update, r.Uuid, r.Policies)
 	if err != nil {
 		return nil, err
 	}
@@ -129,11 +129,11 @@ func (h *Handler) DeleteRole(ctx context.Context, req *idm.DeleteRoleRequest) (*
 	}
 
 	var roles []*idm.Role
-	if err := h.dao(ctx).Search(req.Query, &roles); err != nil {
+	if err := h.dao.Search(req.Query, &roles); err != nil {
 		return nil, err
 	}
 
-	numRows, err := h.dao(ctx).Delete(req.Query)
+	numRows, err := h.dao.Delete(req.Query)
 	if err != nil {
 		return nil, err
 	}
@@ -143,12 +143,12 @@ func (h *Handler) DeleteRole(ctx context.Context, req *idm.DeleteRoleRequest) (*
 
 	for _, r := range roles {
 		// Errors a ignored until now. Should we stop and return an error or better handle the error?
-		err2 := h.dao(ctx).DeletePoliciesForResource(r.Uuid)
+		err2 := h.dao.DeletePoliciesForResource(r.Uuid)
 		if err2 != nil {
 			log.Logger(ctx).Error("could not delete policies for removed role "+r.Label, zap.Error(err2))
 			continue
 		}
-		err2 = h.dao(ctx).DeletePoliciesBySubject(fmt.Sprintf("role:%s", r.Uuid))
+		err2 = h.dao.DeletePoliciesBySubject(fmt.Sprintf("role:%s", r.Uuid))
 		if err2 != nil {
 			log.Logger(ctx).Error("could not delete policies by subject for removed role "+r.Label, zap.Error(err2))
 			continue
@@ -173,7 +173,7 @@ func (h *Handler) SearchRole(request *idm.SearchRoleRequest, response idm.RoleSe
 
 	var roles []*idm.Role
 
-	if err := h.dao(response.Context()).Search(request.Query, &roles); err != nil {
+	if err := h.dao.Search(request.Query, &roles); err != nil {
 		return err
 	}
 
@@ -189,7 +189,7 @@ func (h *Handler) SearchRole(request *idm.SearchRoleRequest, response idm.RoleSe
 // CountRole in database
 func (h *Handler) CountRole(ctx context.Context, request *idm.SearchRoleRequest) (*idm.CountRoleResponse, error) {
 
-	count, err := h.dao(ctx).Count(request.Query)
+	count, err := h.dao.Count(request.Query)
 	if err != nil {
 		return nil, err
 	}
@@ -206,7 +206,7 @@ func (h *Handler) StreamRole(streamer idm.RoleService_StreamRoleServer) error {
 		}
 
 		var roles []*idm.Role
-		if err := h.dao(streamer.Context()).Search(incoming.Query, &roles); err != nil {
+		if err := h.dao.Search(incoming.Query, &roles); err != nil {
 			return err
 		}
 
