@@ -25,6 +25,8 @@ package workspace
 
 import (
 	"context"
+	"github.com/pydio/cells/v4/common/storage"
+	"gorm.io/gorm"
 
 	"github.com/pydio/cells/v4/common/dao"
 	"github.com/pydio/cells/v4/common/sql"
@@ -43,10 +45,14 @@ type DAO interface {
 	Search(sql.Enquirer, *[]interface{}) error
 }
 
-func NewDAO(ctx context.Context, o dao.DAO) (dao.DAO, error) {
-	switch v := o.(type) {
-	case sql.DAO:
-		return &sqlimpl{Handler: v.(*sql.Handler)}, nil
+func NewDAO(ctx context.Context, store storage.Storage) (dao.DAO, error) {
+	var db *gorm.DB
+
+	if store.Get(ctx, &db) {
+		return &sqlimpl{
+			db: db,
+		}, nil
 	}
-	return nil, dao.UnsupportedDriver(o)
+
+	return nil, storage.UnsupportedDriver(store)
 }
