@@ -40,13 +40,13 @@ type Handler struct {
 	idm.UnimplementedWorkspaceServiceServer
 
 	service.Service
-	dao service.ServiceDAOProvider[workspace.DAO]
+	dao workspace.DAO
 }
 
-func NewHandler(ctx context.Context, svc service.Service) idm.WorkspaceServiceServer {
+func NewHandler(ctx context.Context, svc service.Service, dao workspace.DAO) idm.WorkspaceServiceServer {
 	return &Handler{
 		Service: svc,
-		dao:     service.DAOProvider[workspace.DAO](svc),
+		dao:     dao,
 	}
 }
 
@@ -56,10 +56,10 @@ func (h *Handler) CreateWorkspace(ctx context.Context, req *idm.CreateWorkspaceR
 	if req.Workspace.Slug == "" {
 		req.Workspace.Slug = slug.Make(req.Workspace.Label)
 	}
-	update, err := h.dao(ctx).Add(req.Workspace)
+	update, err := h.dao.Add(req.Workspace)
 	// ADD POLICIES
 	if len(req.Workspace.Policies) > 0 {
-		if e := h.dao(ctx).AddPolicies(update, req.Workspace.UUID, req.Workspace.Policies); e != nil {
+		if e := h.dao.AddPolicies(update, req.Workspace.UUID, req.Workspace.Policies); e != nil {
 			return nil, e
 		}
 	}
@@ -101,11 +101,11 @@ func (h *Handler) CreateWorkspace(ctx context.Context, req *idm.CreateWorkspaceR
 func (h *Handler) DeleteWorkspace(ctx context.Context, req *idm.DeleteWorkspaceRequest) (*idm.DeleteWorkspaceResponse, error) {
 
 	workspaces := new([]interface{})
-	if err := h.dao(ctx).Search(req.Query, workspaces); err != nil {
+	if err := h.dao.Search(req.Query, workspaces); err != nil {
 		return nil, err
 	}
 
-	numRows, err := h.dao(ctx).Del(req.Query)
+	numRows, err := h.dao.Del(req.Query)
 	response := &idm.DeleteWorkspaceResponse{
 		RowsDeleted: numRows,
 	}
@@ -116,7 +116,7 @@ func (h *Handler) DeleteWorkspace(ctx context.Context, req *idm.DeleteWorkspaceR
 	// Update relevant policies and propagate event
 	for _, w := range *workspaces {
 		currW := w.(*idm.Workspace)
-		err2 := h.dao(ctx).DeletePoliciesForResource(currW.UUID)
+		err2 := h.dao.DeletePoliciesForResource(currW.UUID)
 		if err2 != nil {
 			log.Logger(ctx).Error("could not delete policies for removed ws "+currW.Slug, zap.Error(err2))
 			continue
@@ -142,13 +142,13 @@ func (h *Handler) SearchWorkspace(request *idm.SearchWorkspaceRequest, response 
 	ctx := response.Context()
 
 	workspaces := new([]interface{})
-	if err := h.dao(ctx).Search(request.Query, workspaces); err != nil {
+	if err := h.dao.Search(request.Query, workspaces); err != nil {
 		return err
 	}
 	var e error
 	for _, in := range *workspaces {
 		ws, ok := in.(*idm.Workspace)
-		if ws.Policies, e = h.dao(ctx).GetPoliciesForResource(ws.UUID); e != nil {
+		if ws.Policies, e = h.dao.GetPoliciesForResource(ws.UUID); e != nil {
 			log.Logger(ctx).Error("cannot load policies for workspace "+ws.UUID, zap.Error(e))
 			continue
 		}
@@ -181,14 +181,14 @@ func (h *Handler) StreamWorkspace(streamer idm.WorkspaceService_StreamWorkspaceS
 		}
 
 		workspaces := new([]interface{})
-		if err := h.dao(ctx).Search(incoming.Query, workspaces); err != nil {
+		if err := h.dao.Search(incoming.Query, workspaces); err != nil {
 			continue
 		}
 
 		var e error
 		for _, in := range *workspaces {
 			if ws, ok := in.(*idm.Workspace); ok {
-				if ws.Policies, e = h.dao(ctx).GetPoliciesForResource(ws.UUID); e != nil {
+				if ws.Policies, e = h.dao.GetPoliciesForResource(ws.UUID); e != nil {
 					log.Logger(ctx).Error("cannot load policies for workspace "+ws.UUID, zap.Error(e))
 					continue
 				}
